+++ quicktill/pricecheck.py
@@ -72,8 +72,8 @@ class pricecheck_continuous_stockline(pricecheck_keypress, ui.menu):
             tillconfig.currency,
             stockline.stocktype.pricestr),
                  "",
-                 "There are {} {}s remaining in stock.".format(
-                     stockline.remaining, stockline.stocktype.unit.name),
+                 "There are {} remaining in stock.".format(
+                     stockline.stocktype.unit.format_qty(stockline.remaining)),
                  "", "Choose a stock item for more information, or "
                  "press another line key.",
         ]
@@ -89,7 +89,7 @@ class pricecheck_stockitem(pricecheck_keypress, ui.listpopup):
     def __init__(self, stockitem):
         td.s.add(stockitem)
         super().__init__(
-            stock.stockinfo_linelist(stockitem.id),
+            [ui.lrline(x) for x in stock.stockinfo_linelist(stockitem.id)],
             title="Stock item {}".format(stockitem.id),
             dismiss=keyboard.K_CASH,
             show_cursor=False,